Split `TestChain` properties in separate modules
based on what they are testing, leaving `TestChain` to contain utility functions
