+++ eras/shelley/test-suite/test/Tests.hs
@@ -7,7 +7,7 @@ import Cardano.Ledger.Shelley.Rules (ShelleyLEDGER)
 import System.IO (hSetEncoding, stdout, utf8)
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (C)
 import Test.Cardano.Ledger.Shelley.Pretty (prettyTest)
-import Test.Cardano.Ledger.Shelley.PropertyTests (depositTests, minimalPropertyTests, propertyTests)
+import Test.Cardano.Ledger.Shelley.PropertyTests (depositsProps, minimalPropertyTests, propertyTests)
 import Test.Cardano.Ledger.Shelley.Rewards (rewardTests)
 import Test.Cardano.Ledger.Shelley.RulesTests (chainExamples, multisigExamples, testTickF)
 import Test.Cardano.Ledger.Shelley.SafeHash (safeHashTest)
@@ -26,7 +26,7 @@ mainTests :: TestTree
 mainTests =
   testGroup
     "Ledger with Delegation"
-    [ depositTests @C
+    [ depositsProps @C
     , minimalPropertyTests @C @(ShelleyLEDGER C)
     , rewardTests
     , Serialisation.tests 5
